敵の名前表示したりダメージ計算途中

+++ Assets/Scripts/ERBehaviours/Action/BeginAttack.cs
@@ -20,7 +20,7 @@ namespace ER.ERBehaviour
         private DirectorWrapMode wrapMode = default;
 
         [SerializeField]
-        private int power = default;
+        private float power = default;
 
         [SerializeField]
         private HandType handType = default;
